Personal Finance Calculator updated

+++ python_introduction/finance_calculator.py
@@ -3,8 +3,7 @@ expenses = int(input("Enter your total monthly expenses: "))
 
 savings = income - expenses
 
-projected_savings = savings * 12 + (savings * 12 * 0.05)
+projected_savings = int(savings * 12 + (savings * 12 * 0.05))
 
 print("Your monthly savings are $",savings)
 print("Projected savings after one year, with interest, is: $",projected_savings)
-
